backend_py/requirements.txt
langchain
python-dotenv
langgraph
langchain-openai
fastapi 
uvicorn
langchain_community
pandas
python-multipart
toml